改成只按键盘相邻的字母有匹配的。键盘距离远的都不管了。

+++ app/src/main/java/com/dataStructure/MetricSpace.java
@@ -6,6 +6,6 @@ package com.dataStructure;
 
 public interface MetricSpace{
 
-    int distance(String a, String b);
+    double distance(String a, String b);
 
 }
